refactor(lut): drop output-neutral double promotions from LUT chain
Scaling by 0.5, 0.75 or 0.1875 and the integer-step sca1 subtraction give
bit-identical results in f32 and f64 (checked exhaustively over all f32
inputs for the multiplications), so keep these in plain f32 arithmetic.
Scene outputs and the C coefficient comparison are unchanged.

+++ lasrc_core/src/atmospheric.rs
@@ -181,7 +181,7 @@ pub fn atmcorlamb2(
         // C: mraot550nm = (raot550nm / normext[indx]) * (pow((lambda[iband] * lambda_sf), -eps));
         // float / float, then * double pow(float * double, -float)
         let base = lambda[iband] as f32 as f64 * lambda_sf;
-        let power = base.powf(-(eps as f32) as f64);
+        let power = base.powf(-eps);
         ((raot550nm as f32 / normext_val) as f64 * power) as f32
     };
